bitbucket/requirements.txt
docker==7.1.0
python-dotenv==1.0.1
pyyaml==6.0.2
hvac==2.3.0
psycopg2-binary==2.9.9
tenacity==9.0.0
requests==2.32.3
prometheus-client==0.21.0